requirements.txt
torch>=1.12.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
networkx>=2.6
rdkit-pypi>=2022.3.1
